Rename controller module to bot module
